Refactor theme application into helper method
