+++ app/src/main/java/ca/pkay/rcloneexplorer/AboutLibsActivity.java
@@ -1,10 +1,6 @@
 package ca.pkay.rcloneexplorer;
 
-import android.app.ActivityManager;
 import android.content.Intent;
-import android.content.SharedPreferences;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.net.Uri;
 import android.os.Bundle;
 import androidx.appcompat.app.ActionBar;
@@ -12,7 +8,6 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.appcompat.widget.Toolbar;
-import androidx.preference.PreferenceManager;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -21,7 +16,7 @@ import java.util.Map;
 
 import ca.pkay.rcloneexplorer.RecyclerViewAdapters.AboutLibrariesAdapter;
 
-import static ca.pkay.rcloneexplorer.StartActivity.tryStartActivity;
+import static ca.pkay.rcloneexplorer.ActivityHelper.tryStartActivity;
 
 public class AboutLibsActivity extends AppCompatActivity implements AboutLibrariesAdapter.OnClickListener {
 
@@ -33,7 +28,7 @@ public class AboutLibsActivity extends AppCompatActivity implements AboutLibrari
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        applyTheme();
+        ActivityHelper.applyTheme(this);
         setContentView(R.layout.activity_about_libs);
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
@@ -56,25 +51,6 @@ public class AboutLibsActivity extends AppCompatActivity implements AboutLibrari
     public boolean onSupportNavigateUp() {
         onBackPressed();
         return true;
-    }
-
-    private void applyTheme() {
-        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
-        int customPrimaryColor = sharedPreferences.getInt(getString(R.string.pref_key_color_primary), -1);
-        int customAccentColor = sharedPreferences.getInt(getString(R.string.pref_key_color_accent), -1);
-        Boolean isDarkTheme = sharedPreferences.getBoolean(getString(R.string.pref_key_dark_theme), false);
-        getTheme().applyStyle(CustomColorHelper.getPrimaryColorTheme(this, customPrimaryColor), true);
-        getTheme().applyStyle(CustomColorHelper.getAccentColorTheme(this, customAccentColor), true);
-        if (isDarkTheme) {
-            getTheme().applyStyle(R.style.DarkTheme, true);
-        } else {
-            getTheme().applyStyle(R.style.LightTheme, true);
-        }
-
-        // set recents app color to the primary color
-        Bitmap bm = BitmapFactory.decodeResource(getResources(), R.mipmap.ic_launcher_round);
-        ActivityManager.TaskDescription taskDesc = new ActivityManager.TaskDescription(getString(R.string.app_name), bm, customPrimaryColor);
-        setTaskDescription(taskDesc);
     }
 
     private void createData() {
